requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
Pillow==11.0.0

